requirements.txt
Click
Jinja2
six
OWSLib
pyyaml
